requirements.txt
fastapi
uvicorn
psycopg2-binary
langchain
langchain_community
python-dotenv
